Cleaned up a few items
Gave the exploding slimeball the same death anim as the slimeball and
the sailor. Will probably make that the case for all enemies.

Resized the slimeball's canvas to make it the same size as the death
anim

Set sailor and slimeball sprites origin to middle center and updated
collision checks accordingly.